requirements.txt
openai>=1.35.0
python-dotenv
